BLC-508 - Remove price aspect from the framework
Removing the aspect exposed some issues with cart operations that return an OrderItem.     Because the pricing service may combine or split order items, in some cases, the system was returning orphaned items.

Added a new attribute to orderService that will prevent automatic merging of cart items.   Items will still be split (for example in BOGO) situations.