Check property flag to determine if new Sitemap.xml should be generated after timeout.
The default implementation will first check local resources and then retrieve from file system if timeout is expired. In the default implementation; the only way to generate a new sitemap.xml once it has been created is to delete the current sitemap.xml from the file system. This new flag will provide an option to override that behavior and allow a new sitemap.xml to be created after timeout instead of returning an old (potentially outdated) sitemap.xml from the file system.

- Addresses BroadleafCommerce/QA#941