BroadleafCommerce/QA#1880 - Validate URL and param without ESAPI When switching between sandboxes, ESAPI would throw a validation error for the redirect produced because it is not subordinate to `/admin` or `/login`, and ESAPI does not let you use a full URL to validate redirects, only relative URLs. Therefore, I moved away from ESAPI in this instance.
Instead, I take the request base URL and compare that to the site base url to make sure they are the same. We do not want to be redirected to some other domain or a malicious page. (By default, when you change sandboxes, you are sent back to the index page (i.e., `/`).) If not valid, then it is a security error. After that check, if the URL param "blSandboxId" is present, we check to make sure that the sandbox it references actually exists. We also ignore all other parameters, because there should not be any.