#812 - Move all of the admin beans that are created in the site context to only be created when running the admin (move from content-client to content-creator)
This also removes some beans that were not being used at all and changed some unnecessarily-prototype-scoped beans into singletons. One of the changes is that DynamicEntityDao and its ilk is never available in the site context. This actually shouldn't have happened to begin with as there was a snowball effect of other admin-only class dependencies (like FieldMetadataProviders) that should never be instantiated in the site