#850 - Rewrite the Broadleaf property configurations to utilize an ApplicationListener and spring.factories entries
The main reason to do this is because registering PropertySources with a BeanFactoryPostProcessor means that we cannot use these properties to drive @ConditionalOnProperty, since they weren't registered early enough. An ApplicationListener is activated immediately after the ApplicationContext has been created and it is a well-defined hookpoint in various places within Spring.

This did require some changes in how we do test configuration (since we
need eo specify the initializerClasses now in the
@ContextConfiguration). This also means that since @PropertySource
annotations are not determined until the ConfigurationClassParser runs
the framework properties will take precedence over any properties
defined there. If someone is using Spring Boot though, anything in
application.properties overrides any of the Broadleaf properties.

Since this is the case, I have also changed how the PropertySources are
registered within the Spring Environment. Previously every
PropertySource from the framework or profile was registered individually
in the Environment. This is now a CompositePropertySource for profile
and framework properties so that you can reorder or change all of them
as a single unit.

This adds some exploratory testing to validate the above assumptions of
how things work.