Shore up integration testing configuration (#869)
This adds a few new annotations that can be used for integration tests in TestNG, JUnit and Spock contexts. This also greatly simplifies the configuration between the 3 since they now all can depend on the same base annotations rather than copy/pasting the same configuration around to the different scenarios. For JUnit specifically, we have removed the base classes and instead rely on annotation composition. If a particular user decides to extend from one of the Spring base classes then they can but it is no longer required. TestNG tests still require extension from one of the Broadleaf base classes because they are tied to the Spring TestNG class hierarchy. However, this is now split between transactional and non-default-transactional tests