Add the ability to read index fields by their abbreviation and entity type (#933)
Generally, an abbreviation should be unique by its abbreviation within a particular document. But since there are multiple document types, it's possible that you can have the same abbreviation but within different entity types. Therefore, it makes sense to specify it on a per-type basis.

This was done in response to fixing an issue from BroadleafCommerce/QA#2791