Retrieved the SKU entity instead of just the ID because we need to manually filter out archived SKUs before returning their IDs.
The normal filters that would handle this automagically for queries only apply to the root entity (SkuProductOptionValueXref here). However, for MT cases, we also need to manually make sure that the SKU doesn't have a local override that's archived. Thus, by retrieving the entity instead of the ID, we'll get the entity that's already been transformed to the correct sandbox version for the current site, and we can check that for archived status before returning the correct IDs.

see BroadleafCommerce/QA#2825