Added new cache to store dates by URI for when queries are made when looking up a Page by URI.
This will reduce the number of queries made in this context from 1 per URI per 10 minutes, to 1 per URI per day, unless the cmsPageCache is invalidated due to sandbox changes.

`findPageBy()` looks up pages that are active and online at the **date** of the request (the **date** is cached for 10 minutes). These changes will change that interval to **1 day** and account for potential cache invalidations due to changes to the production page (deploys) by checking the **cmsPageCache** before checking the query date cache. The new method, `findPageByURIAndActiveDate()`, retrieves Pages that are active at the time of the cached date and up to a day later. Because we cache the query for a day, we need to retrieve all pages that will be active any time during that interval even if they aren't active for the entire interval. Therefore, we then need to manually remove any pages that aren't active right now.

see BroadleafCommerce/QA#2727