Add a ServletContextListener that can correctly bootstrap a Spring Boot application from a Servlet 3.0 web.xml (#977)
This was done to support versions of Tomcat 7 (and perhaps other servlet containers) whose ServletContextInitializer scanning loads and then caches classes prior to Broadleaf adding requisite class transformers. If this scanning needs to be disabled for this reason, you have to bootstrap from a web.xml. This is the simplest configuration that plays nicely with how Spring Boot would normally bootstrap the application via a ServletContextInitializer