BroadleafCommerce/QA#2948 - Clone policy is required fro the structured content field
Since we programmatically modify some of the ToOne properties on the StructuredContentFieldXref to the StructuredContentField entity, we must indicate to Broadleaf that it should be cloned. If not, this fails silently and simply does not process any of the fields we set on the StructuredContentField at all.

In order to expose these fields to populate in the first place, we need to have populateToOneFields set to true