#1887 Reimplemented the old functionality for translating a hibernate cache key to our hydrated cache
CacheKey was removed in Hibernate 5 and replaced with by CacheKeyImplementation. However CacheKeyImplementaiton is a protected class and no longer has the same methods as CacheKey. In order to get around this we test to see if we're caching a hibernate enitty by directly checking if the class name of the key given is the class name of CacheKeyImplementation. Additionally in order to break the key up like we did before we utilize that the toString is built of the pieces we needed previously therefore we just split the string on the "#" since the part before it is the entity and the part after it is the id