Fixes issue where if you select an address that has a tax and then switch to an address that has a tax that the tax for the original address is still applied
The issue was caused by the conditional checking if the tax was zero
where we would entirely skip handing any tax details. Because of this if
there was already a non-zero tax applied it wouldn't be removed when the
tax was changed to needing to be zero. The core logic around this was
moved to applyTaxFactor where now it handles looking for the case where
if a tax record is found and the tax factor to be applied is zero then
the tax record is removed from the list of tax details.